Roman_Urdu_Conversion_Project/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
transformers>=4.15.0
scikit-learn>=1.0.0
nltk>=3.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
tqdm>=4.62.0
urdu-digit>=0.1.0
arabic-reshaper>=2.1.0
python-bidi>=0.4.2
